fix: do not send trasnaction if blob already certified, tidy publisher traces (#2519)
When sending a request to the publisher to store a blob that was already stored, certification
would be skipped but a transaction would be sent to Sui regardless due to not checking if there
were no blobs to put into the PTB. This fixes that skips several code paths when the list on which
they would be operating is empty.

Additionally, this tidies some of the tracing export to reduce noise in trace exports.